requirements-kg2-build.txt
argh==0.26.2
cachier==1.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
conf==0.4.1
dataclasses==0.6
decorator==4.4.0
HTMLParser==0.0.2
idna==2.8
isodate==0.6.0
jsobject==0.10.2
marshmallow==3.0.0b3
mysqlclient==1.4.4
neo4j==1.7.4
networkx==2.2
numpy==1.16.3
ontobio==1.10.1
pandas==0.24.2
pathtools==0.1.2
portalocker==1.4.0
prefixcommons==0.1.7
pymongo==3.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
pysolr==3.8.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
rdflib==4.2.2
requests==2.21.0
scipy==1.2.1
six==1.12.0
snakemake==5.5.4
SPARQLWrapper==1.8.4
urllib3==1.24.3
watchdog==0.9.0
xmltodict==0.12.0
yamldown==0.1.8
validators==0.15.0
